+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [#90] Capture HTTP response headers in django and flask.
 
+### Changed
+- Headers such as Host, User-Agent, Authorization and Content-Type are no longer
+  filtered out in HTTP event `headers` field.
+
 ## [0.9.0] - 2021-05-03
 ### Added
 - Attempting to start recording while recording is in progress will now raise a